+++ network/p2p/p2pbuilder/libp2pNodeBuilder.go
@@ -1,4 +1,4 @@
-package builder
+package p2pbuilder
 
 import (
 	"context"
